Refactor ParseFieldOperations
